refactor: move the logic to swap between API and resource into separate package
