+++ internal/provider/provider.go
@@ -13,6 +13,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/kibblator/terraform-provider-ory/internal/provider/resources"
+	"github.com/kibblator/terraform-provider-ory/internal/provider/resources/email_configuration_resource"
 	orytypes "github.com/kibblator/terraform-provider-ory/internal/provider/types"
 
 	openapiclient "github.com/ory/client-go"
@@ -217,6 +218,6 @@ func (p *oryProvider) DataSources(_ context.Context) []func() datasource.DataSou
 func (p *oryProvider) Resources(_ context.Context) []func() resource.Resource {
 	return []func() resource.Resource{
 		resources.NewRegistrationResource,
-		resources.NewEmailConfigurationResource,
+		email_configuration_resource.NewEmailConfigurationResource,
 	}
 }
